+++ xdef/changelog.md
@@ -1,7 +1,8 @@
 # Version ${version}, release-date ${release.date}
 
-# Version 40.0.1, release-date 2020-04-25
-* improved JSON error reporting
+# Version 40.0.1, release-date 2020-04-30
+* improved JSON error reporting.
+* corrected bug when a JSON string value contains some of escape characters.
 
 # Version 40.0.0, release-date 2020-04-24
 * new implementation of processing of JSON data.